requirements.txt
playwright>=1.40.0
requests>=2.31.0
pynacl>=1.5.0
pyyaml>=6.0
